Move all handlers from ext to core. These handlers only have dependencies on Java SE.
