+++ core/src/main/java/org/jboss/logmanager/handlers/QueueHandler.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.ext.handlers;
+package org.jboss.logmanager.handlers;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
