+++ core/src/main/java/org/jboss/logmanager/handlers/TcpOutputStream.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.ext.handlers;
+package org.jboss.logmanager.handlers;
 
 import java.io.Closeable;
 import java.io.Flushable;
