+++ core/src/test/java/org/jboss/logmanager/handlers/AssertingErrorManager.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package org.jboss.logmanager.ext.handlers;
+package org.jboss.logmanager.handlers;
 
 import java.io.IOException;
 import java.io.PrintWriter;
