+++ core/src/test/resources/client-keystore.jks
