+++ core/src/test/resources/client.cer
