+++ core/src/test/resources/generate.sh
