+++ core/src/test/resources/server-keystore.jks
