+++ core/src/test/resources/server.cer
