+++ core/src/test/resources/test-client-store.jks
